+++ src/Firebase/Auth/SignInResult.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Auth;
 
+use Kreait\Firebase\JWT\Token\Parser;
 use Lcobucci\JWT\Encoding\JoseEncoder;
-use Lcobucci\JWT\Token\Parser;
 use Lcobucci\JWT\UnencryptedToken;
 use stdClass;